+++ src/Stepper/test/storySettings.js
@@ -1,7 +1,7 @@
-import { Category } from '../../../../stories/storiesHierarchy';
+import { Category } from '../../../stories/storiesHierarchy';
 
 export const storySettings = {
-  category: Category.BETA,
+  category: Category.WIP,
   storyName: 'Stepper',
   dataHook: 'story-stepper',
   testStoryNames: {
